backend/requirements.txt

fastapi
uvicorn
boto3
sqlalchemy
psycopg2-binary
python-multipart
fpdf
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-http